Add code to #create action
The create action uses a helper method to obtain the title and body data
when creating a new post. It also sets the id using the helper method
current_user from devise. We check if the post is saved. When it's true,
we redirect to index view. Else, we render the same page and send an
unprocessable entity response.