Create show and index views
Added route to show action.

Also added code to index and show actions.

Inside the show view, there's code to display the post title and body.
And for index, it shows a list of the posts titles, and when a user is
signed, it also shows who wrote each post.

Lastly, added associations to both post and user models.