Add flash messages to Post controller
Added flash messages when the user tries to creates a post but fails to
create it. Also made a partial to render this flash in the index and new
views.

Also added a condition to the sign_out partial to display the sign out
link only when a user is signed.